==> R/00_all.qmd <==
---
title: "0_all"
format: html
editor: visual
---

## Load Libraries

```{r}
#Check if libraries have to be installed
rm(list=ls()) # remove all variables in the environment
packages <- c("here", "tidyverse")
install.packages(base::setdiff(packages, rownames(installed.packages()))) 


# Load required library
library("here")
```

```{r}
# List of all the files created
files <- c("R/01_load_and_tidy.qmd",
           "R/02_augment_of_count_data.qmd",
           "R/03_augment_and_join_of_data_frames.qmd",
           "R/04_explorative_data_analysis.qmd",
           "R/05_explorative_analysis_and_DGE_analysis_on_normalized_counts_using_glm.qmd",
           "R/06_DGE_within_all_tissues.qmd",
           "R/07_Data_analysis_historgram_and_freq.qmd")

# Render all files
lapply(files, function(file) {
  rmarkdown::render(here(file), output_dir = here("results"))
})

```
